Move common bindings defs to pennylane_lightning/src/bindings/BindingsBase.hpp.
