+++ pennylane_lightning/src/bindings/CMakeLists.txt
@@ -1,0 +1,11 @@
+cmake_minimum_required(VERSION 3.20)
+
+project(lightning_bindings LANGUAGES CXX)
+
+add_library(lightning_bindings INTERFACE)
+
+target_include_directories(lightning_bindings INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
+
+target_link_libraries(lightning_bindings INTERFACE lightning_utils)
+
+set_property(TARGET lightning_bindings PROPERTY POSITION_INDEPENDENT_CODE ON)
